+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changes
 
 ## 🔧Fixes🔧
+- Installer not detecting python
 
 ## ⏫Improvements⏫
+- Installer logging
 
 ## ✨New✨